Customize unlock routine
