+++ Passepartout/App/Constants/Theme.swift
@@ -25,6 +25,7 @@
 
 import SwiftUI
 import PassepartoutLibrary
+import LocalAuthentication
 
 extension View {
     var themeIdiom: UIUserInterfaceIdiom {
@@ -491,17 +492,41 @@ extension View {
             EmptyView()
         }
     }
+}
 
+// MARK: Lock screen
+
+extension View {
     func themeLockScreen() -> some View {
         @AppStorage(AppPreference.locksInBackground.rawValue) var locksInBackground = false
         return LockableView(
-            reason: L10n.Global.Messages.unlockApp,
             locksInBackground: $locksInBackground,
             content: {
                 self
             },
-            lockedContent: LogoView.init
+            lockedContent: LogoView.init,
+            unlockBlock: Self.themeUnlockScreenBlock
         )
+    }
+
+    private static func themeUnlockScreenBlock(isLocked: Binding<Bool>) {
+        let context = LAContext()
+        let policy: LAPolicy = .deviceOwnerAuthentication
+        var error: NSError?
+        guard context.canEvaluatePolicy(policy, error: &error) else {
+            isLocked.wrappedValue = false
+            return
+        }
+        Task { @MainActor in
+            do {
+                let isAuthorized = try await context.evaluatePolicy(
+                    policy,
+                    localizedReason: L10n.Global.Messages.unlockApp
+                )
+                isLocked.wrappedValue = !isAuthorized
+            } catch {
+            }
+        }
     }
 }
